requirements.txt
greenlet==3.0.3
paho-mqtt==2.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
typing_extensions==4.11.0
